+++ src/Application.php
@@ -282,30 +282,6 @@ abstract class Application {
 
 
     /**
-     * Define rocket theme as default theme.
-     */
-    public function defineCache()
-    {
-        $cache = get_option('cache');
-        $cache_options = $this->config->get('cache');
-
-        // Cache_Enabler options
-        // expires, new_post, new_comment, compress, webp, excl_ids, minify_html
-
-        if( defined('WP_CACHE') and WP_CACHE and $cache_options and class_exists('Cache_Enabler') )
-        {
-            if( isset($cache_options['http']) and $cache_options['http'] != $cache['expires'] )
-            {
-                $cache_options['expires'] = $cache_options['http']/3600;
-                $cache_options['new_post'] = 1;
-                $cache = array_merge($cache, $cache_options);
-                update_option('cache', $cache);
-            }
-        }
-    }
-
-
-    /**
      * Clean WP Head
      */
     public function afterSetupTheme()
@@ -805,8 +781,6 @@ abstract class Application {
                 echo implode('<br/>', $notices );
             });
         }
-
-        $this->defineCache();
     }