+++ doc/changelog.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+# 1.2.32 - 2017-12-11
+### Removed
+* Cache plugin
+### Fixed
+* $base_uri on ssl
+
 # 1.2.31 - 2017-11-20
 ### Added
 * Query::wp_query